requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --index-url=https://mirrors.aliyun.com/pypi/simple/ --no-emit-index-url requirements/dev.in
#
aerich==0.2.3             # via -r requirements/././production.in
aiosqlite==0.15.0         # via tortoise-orm
anyio==2.0.0              # via asyncclick
appdirs==1.4.4            # via black, virtualenv
appnope==0.1.0            # via ipython
argon2-cffi==20.1.0       # via -r requirements/././production.in
astroid==2.4.1            # via prospector, pylint, pylint-celery, pylint-flask, requirements-detector
asyncclick==7.1.2.1       # via aerich
asyncpg==0.21.0           # via -r requirements/././production.in
asynctest==0.13.0         # via -r requirements/./test.in
attrs==20.2.0             # via pytest
backcall==0.2.0           # via ipython
black==20.8b1             # via -r requirements/dev.in
certifi==2020.6.20        # via httpx, requests, sentry-sdk
cffi==1.14.3              # via argon2-cffi
cfgv==3.2.0               # via pre-commit
chardet==3.0.4            # via requests
click==7.1.2              # via black, pip-tools, uvicorn
coverage==5.3             # via -r requirements/dev.in, pytest-cov
decorator==4.4.2          # via ipython
distlib==0.3.1            # via virtualenv
dodgy==0.2.1              # via prospector
faker==4.1.3              # via -r requirements/./test.in
fastapi==0.61.1           # via -r requirements/././production.in
filelock==3.0.12          # via virtualenv
flake8-polyfill==1.0.2    # via pep8-naming
flake8==3.8.3             # via flake8-polyfill
gunicorn==20.0.4          # via -r requirements/././production.in
h11==0.9.0                # via httpcore, uvicorn
httpcore==0.11.0          # via httpx
httptools==0.1.1          # via uvicorn
httpx==0.15.2             # via -r requirements/././production.in
identify==1.5.4           # via pre-commit
idna==2.10                # via anyio, requests, rfc3986
iniconfig==1.0.1          # via pytest
ipython-genutils==0.2.0   # via traitlets
ipython==7.18.1           # via -r requirements/dev.in
iso8601==0.1.13           # via tortoise-orm
isort==4.3.21             # via pylint
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via -r requirements/././production.in
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8, prospector, pylint
more-itertools==8.5.0     # via pytest
mypy-extensions==0.4.3    # via black
nodeenv==1.5.0            # via pre-commit
packaging==20.4           # via pytest, pytest-sugar
parso==0.7.1              # via jedi
pathspec==0.8.0           # via black
pep8-naming==0.10.0       # via prospector
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.3.1          # via -r requirements/dev.in
pluggy==0.13.1            # via pytest
pre-commit==2.7.1         # via -r requirements/dev.in
prompt-toolkit==3.0.7     # via ipython
prospector==1.3.0         # via -r requirements/dev.in
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via flake8, prospector
pycparser==2.20           # via cffi
pydantic==1.6.1           # via -r requirements/././production.in, aerich, fastapi
pydocstyle==5.1.1         # via prospector
pyflakes==2.2.0           # via flake8, prospector
pygments==2.7.1           # via ipython
pylint-celery==0.3        # via prospector
pylint-django==2.0.15     # via prospector
pylint-flask==0.6         # via prospector
pylint-plugin-utils==0.6  # via prospector, pylint-celery, pylint-django, pylint-flask
pylint==2.5.2             # via prospector, pylint-celery, pylint-django, pylint-flask, pylint-plugin-utils
pyparsing==2.4.7          # via packaging
pypika==0.39.1            # via tortoise-orm
pytest-cov==2.10.1        # via -r requirements/./test.in
pytest-sugar==0.9.4       # via -r requirements/./test.in
pytest==6.0.2             # via -r requirements/./test.in, pytest-cov, pytest-sugar
python-dateutil==2.8.1    # via faker
python-multipart==0.0.5   # via -r requirements/././production.in
pyyaml==5.3.1             # via pre-commit, prospector
regex==2020.7.14          # via black
requests==2.24.0          # via -r requirements/././production.in
requirements-detector==0.7  # via prospector
rfc3986[idna2008]==1.4.0  # via httpx
sentry-sdk==0.17.8        # via -r requirements/././production.in
setoptconf==0.2.0         # via prospector
six==1.15.0               # via argon2-cffi, astroid, packaging, pip-tools, python-dateutil, python-multipart, virtualenv
sniffio==1.1.0            # via anyio, httpcore, httpx
snowballstemmer==2.0.0    # via pydocstyle
starlette==0.13.6         # via fastapi
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.3       # via faker
toml==0.10.1              # via black, pre-commit, pylint, pytest
tortoise-orm==0.16.15     # via -r requirements/././production.in, aerich
traitlets==5.0.4          # via ipython
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via aiosqlite, black, tortoise-orm
urllib3==1.25.10          # via requests, sentry-sdk
uvicorn==0.11.8           # via -r requirements/././production.in
uvloop==0.14.0            # via uvicorn
virtualenv==20.0.31       # via pre-commit
wcwidth==0.2.5            # via prompt-toolkit
websockets==8.1           # via uvicorn
wrapt==1.12.1             # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
